move drawer context in ts
